+++ source_gen/source_gen/lib/source_gen.dart
@@ -10,4 +10,5 @@ export 'src/generator.dart';
 export 'src/generator_for_annotation.dart';
 export 'src/library.dart' show LibraryReader;
 export 'src/revive.dart' show Revivable;
+export 'src/span_for_element.dart' show spanForElement;
 export 'src/type_checker.dart' show TypeChecker;